# openapi/control.openapi.yaml
openapi: 3.0.3
info:
  title: Termix Control API
  version: "1.0.0"
servers:
  - url: /api/v1
paths:
  /auth/login:
    post:
      operationId: postAuthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/refresh:
    post:
      operationId: postAuthRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        "200":
          description: refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        "401":
          description: refresh token invalid or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      operationId: postAuthLogout
      tags: [auth]
      security: []
      responses:
        '204':
          description: logged out
        '400':
          description: no refresh credential present
  /host/sessions:
    post:
      operationId: postHostSessions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        "200":
          description: session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
  /host/sessions/{session_id}/heartbeat:
    post:
      operationId: postHostSessionHeartbeat
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionHeartbeatRequest'
      responses:
        "200":
          description: session heartbeat accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        "401":
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /host/sessions/{session_id}:
    patch:
      operationId: patchHostSession
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionRequest'
      responses:
        "200":
          description: session updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
  /sessions/{session_id}:
    get:
      operationId: getSession
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        "404":
          description: session not found
  /sessions:
    get:
      operationId: listSessions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [running, idle, disconnected, exited, failed, all]
      responses:
        "200":
          description: sessions for the authenticated user
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items: { $ref: '#/components/schemas/Session' }
  /sessions/{session_id}/control/acquire:
    post:
      operationId: postSessionControlAcquire
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: control lease acquired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlLeaseResponse'
        "400":
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: session or device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: lease conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{session_id}/control/renew:
    post:
      operationId: postSessionControlRenew
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlLeaseRequest'
      responses:
        "200":
          description: control lease renewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlLeaseResponse'
        "400":
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: session or device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: lease conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{session_id}/control/release:
    post:
      operationId: postSessionControlRelease
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlLeaseRequest'
      responses:
        "200":
          description: control lease released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseControlLeaseResponse'
        "400":
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: session or device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: lease conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password, device_type, platform, device_label]
      properties:
        email: { type: string, format: email }
        password: { type: string }
        device_type: { type: string, enum: [host, android, ios, web] }
        platform: { type: string, enum: [macos, ubuntu, windows, android, ios, web] }
        device_label: { type: string }
        cookie_mode:
          type: boolean
          default: false
          description: when true, refresh_token is returned as HttpOnly cookie instead of body
    LoginResponse:
      type: object
      required: [user, device, access_token, expires_in_seconds]
      properties:
        user: { $ref: '#/components/schemas/User' }
        device: { $ref: '#/components/schemas/Device' }
        access_token: { type: string }
        refresh_token:
          type: string
          nullable: true
          description: omitted when cookie_mode=true (refresh sent as HttpOnly cookie)
        expires_in_seconds: { type: integer }
        cookie_mode:
          type: boolean
          description: echoed from request for client convenience
    User:
      type: object
      required: [id, email, display_name, role]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { type: string, enum: [admin, user] }
    Device:
      type: object
      required: [id, device_type, platform, label]
      properties:
        id: { type: string, format: uuid }
        device_type: { type: string, enum: [host, android, ios, web] }
        platform: { type: string, enum: [macos, ubuntu, windows, android, ios, web] }
        label: { type: string }
    CreateSessionRequest:
      type: object
      required: [device_id, tool, launch_command, cwd, cwd_label, hostname]
      properties:
        device_id: { type: string, format: uuid }
        tool: { type: string, enum: [claude, codex, opencode] }
        name: { type: string }
        launch_command: { type: string }
        cwd: { type: string }
        cwd_label: { type: string }
        hostname: { type: string }
    CreateSessionResponse:
      type: object
      required: [session_id, tmux_session_name, status]
      properties:
        session_id: { type: string, format: uuid }
        tmux_session_name: { type: string }
        status: { type: string }
    UpdateSessionRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [starting, running, idle, disconnected, exited, failed] }
        last_error: { type: string, nullable: true }
        last_exit_code: { type: integer, nullable: true }
    SessionHeartbeatRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [starting, running, idle] }
    Session:
      type: object
      required: [id, user_id, host_device_id, tool, launch_command, cwd, cwd_label, tmux_session_name, status]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        host_device_id: { type: string, format: uuid }
        name: { type: string, nullable: true }
        tool: { type: string, enum: [claude, codex, opencode] }
        launch_command: { type: string }
        cwd: { type: string }
        cwd_label: { type: string }
        tmux_session_name: { type: string }
        status: { type: string }
        host_label:
          type: string
          nullable: true
          description: human-readable host description (platform plus device label) for the session's host device
        control:
          $ref: '#/components/schemas/SessionControlState'
    SessionControlState:
      type: object
      required: [holder]
      description: |
        Active control-lease state visible to the requesting device. Omitted when no
        active lease exists or the endpoint does not surface lease state. `holder=self`
        means the requesting device currently holds the lease; `holder=other` means a
        different device of the same user holds it.
      properties:
        holder:
          type: string
          enum: [self, other]
        holder_label:
          type: string
          description: human-readable holder description (platform plus device label) when known
    ControlLeaseRequest:
      type: object
      required: [lease_version]
      properties:
        lease_version: { type: integer, format: int64 }
    ControlLeaseResponse:
      type: object
      required: [session_id, controller_device_id, lease_version, granted_at, expires_at, renew_after_seconds]
      properties:
        session_id: { type: string, format: uuid }
        controller_device_id: { type: string, format: uuid }
        lease_version: { type: integer, format: int64 }
        granted_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        renew_after_seconds: { type: integer }
    ReleaseControlLeaseResponse:
      type: object
      required: [session_id, lease_version, released]
      properties:
        session_id: { type: string, format: uuid }
        lease_version: { type: integer, format: int64 }
        released: { type: boolean }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        reason: { type: string }
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          nullable: true
          description: omitted when caller relies on HttpOnly cookie
    RefreshResponse:
      type: object
      required: [access_token, expires_in_seconds]
      properties:
        access_token:        { type: string }
        expires_in_seconds:  { type: integer }
        refresh_token:       { type: string, nullable: true }
        user:
          allOf:
            - $ref: '#/components/schemas/User'
          nullable: true
        device:
          allOf:
            - $ref: '#/components/schemas/Device'
          nullable: true